monitor_app/back-end/pyproject.toml
[project]
name = "back-end"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "kubernetes>=32.0.1",
    "pyhumps>=3.8.0",
]
